rbac spring security is final

+++ src/main/java/az/et/ws/security/JWTUtil.java
@@ -4,7 +4,6 @@ import az.et.ws.component.entity.TokenEntity;
 import az.et.ws.component.exception.InvalidTokenException;
 import az.et.ws.component.mapper.ObjectMapper;
 import az.et.ws.component.model.AppUser;
-import az.et.ws.component.model.Token;
 import az.et.ws.component.response.AuthResponse;
 import az.et.ws.repository.redis.TokenRepository;
 import com.auth0.jwt.JWT;
@@ -18,6 +17,7 @@ import org.springframework.stereotype.Component;
 
 import javax.annotation.PostConstruct;
 import java.util.Date;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -38,25 +38,33 @@ public class JWTUtil {
     public AuthResponse createTokenAndSession(Authentication authentication) {
         AppUser appUser = (AppUser) authentication.getPrincipal();
         AuthResponse authResponse = AuthResponse.builder()
-                .token(generateToken(appUser))
+                .userId(appUser.getId())
+                .username(appUser.getUsername())
+                .accessToken(accessToken(appUser))
+                .refreshToken(refreshToken(appUser))
+                .authorities(appUser.getAuthorities())
                 .build();
-        tokenRepository.save(objectMapper.m2e(authResponse.getToken()));
+        tokenRepository.save(objectMapper.authResponseToCache(authResponse));
         return authResponse;
     }
 
-    public String verifyToken(String authorizationHeader) throws InvalidTokenException {
+    public TokenEntity verifyToken(String authorizationHeader) throws InvalidTokenException {
         String token = authorizationHeader.substring("Bearer ".length());
         if (findAuthToken(token).isPresent()) {
-            return findAuthToken(token).get().getAccessToken();
+            return findAuthToken(token).get();
         } else {
             throw new InvalidTokenException();
         }
     }
 
     public void invalidateToken(String authorizationHeader) throws InvalidTokenException {
-        String token = authorizationHeader.substring("Bearer ".length());
-        if (findAuthToken(token).isPresent()) {
-            tokenRepository.deleteById(token);
+        if (Objects.nonNull(authorizationHeader)) {
+            String token = authorizationHeader.substring("Bearer ".length());
+            if (findAuthToken(token).isPresent()) {
+                tokenRepository.deleteById(token);
+            } else {
+                throw new InvalidTokenException();
+            }
         } else {
             throw new InvalidTokenException();
         }
@@ -80,10 +88,6 @@ public class JWTUtil {
                 .withExpiresAt(new Date(System.currentTimeMillis() + 30 * 60 * 1000))
                 .withClaim("roles", user.getAuthorities().stream().map(GrantedAuthority::getAuthority).collect(Collectors.toList()))
                 .sign(Algorithm.HMAC256(getJwtSecret()));
-    }
-
-    private Token generateToken(AppUser appUser) {
-        return Token.builder().accessToken(accessToken(appUser)).refreshToken(refreshToken(appUser)).build();
     }
 
 }
